requirements.txt
click==8.1.5
colorama==0.4.6
gitdb==4.0.10
GitPython==3.1.32
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.15.1
rich==13.4.2
shellingham==1.5.0.post1
smmap==5.0.0
typer==0.9.0
typing_extensions==4.7.1